The project structure is organized and Bootstrap is integrated
