+++ Ulimaps/src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import reactLogo from '/src/assets/react.svg'
+import viteLogo from '/public/vite.svg'
+import './styles/App.css'
 
 function App() {
   const [count, setCount] = useState(0)
